feat: support for opening login modal after login session expiration
