+++ src/stores/user.ts
@@ -5,12 +5,14 @@ import { defineStore } from "pinia";
 interface UserStoreState {
   currentUser?: User;
   isAnonymous: boolean;
+  loginModalVisible: boolean;
 }
 
 export const useUserStore = defineStore("user", {
   state: (): UserStoreState => ({
     currentUser: undefined,
     isAnonymous: true,
+    loginModalVisible: false,
   }),
   actions: {
     async fetchCurrentUser() {
